+++ tools.go
@@ -4,6 +4,6 @@
 package plugin
 
 import (
-	_ "github.com/golangci/golangci-lint/cmd/golangci-lint"
+	_ "github.com/magefile/mage"
 	_ "gotest.tools/gotestsum"
 )